Generalize Leaper pieces (for Knight only)
