+++ src/chess/pieces/leapers/mod.rs
@@ -1,0 +1,3 @@
+pub mod knight;
+pub mod king;
+pub mod leaper;
